move SearchPage into own component folder for organizationa and normalized file structure
