+++ .env.example
@@ -1,4 +1,4 @@
 FLASK_APP=app
 FLASK_ENV=development
 SECRET_KEY=lkasjdf09ajsdkfljalsiorj12n3490re9485309irefvn,u90818734902139489230
-DATABASE_URL=postgresql://starter_app_dev@localhost/starter_app
+DATABASE_URL=postgresql://starter_app:password@localhost/starter_app_dev
