+++ react-app/src/components/SearchPage/search_page.css
